Create servernuker.lua (PATCHED)
One of the most destructive scripts ever, a servernuker.

PSA: Patched in Da Hood.

Like Smite, it instantly kills people in a Da Hood server. 
What made this different, was the fact that instead of killing one person it duplicated 40+ spray cans and killed the whole server.

Made for some pretty funny TikToks of people watching the whole server get teleported to the void.